requirements.txt
aiogram
python-telegram-bot
pandas
pybit
python-dotenv
aiohttp
matplotlib
numpy